Moved the tense guide to a separate file
